requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pandas>=1.2.0
pyyaml>=5.4.0

# Image processing
Pillow>=8.0.0
aggdraw>=1.3.11

# Geometry and shapes
shapely>=1.7.0

# NLP
nltk>=3.6

# Machine learning utilities
scikit-learn>=0.24.0

# Progress bars
tqdm>=4.60.0

# For vision models (if using pretrained)
torchsummary>=1.5.1

# Visualization (optional, for debugging)
matplotlib>=3.3.0

